fix: ErrorBoundary in console
